10.3.9

+++ cctools/include/mach-o/swap.h
@@ -23,6 +23,7 @@
 #ifndef _MACH_O_SWAP_H_
 #define _MACH_O_SWAP_H_
 
+#include <stdint.h>
 #include <architecture/byte_order.h>
 #include <mach-o/fat.h>
 #include <mach-o/loader.h>
@@ -36,7 +37,7 @@ extern void swap_fat_header(
 
 extern void swap_fat_arch(
     struct fat_arch *fat_archs,
-    unsigned long nfat_arch,
+    uint32_t nfat_arch,
     enum NXByteOrder target_byte_order);
 
 extern void swap_mach_header(
@@ -53,7 +54,7 @@ extern void swap_segment_command(
 
 extern void swap_section(
     struct section *s,
-    unsigned long nsects,
+    uint32_t nsects,
     enum NXByteOrder target_byte_order);
 
 extern void swap_symtab_command(
@@ -126,42 +127,42 @@ extern void swap_prebind_cksum_command(
 
 extern void swap_twolevel_hint(
     struct twolevel_hint *hints,
-    unsigned long nhints,
+    uint32_t nhints,
     enum NXByteOrder target_byte_sex);
 
 extern void swap_nlist(
     struct nlist *symbols,
-    unsigned long nsymbols,
+    uint32_t nsymbols,
     enum NXByteOrder target_byte_order);
 
 extern void swap_ranlib(
     struct ranlib *ranlibs,
-    unsigned long nranlibs,
+    uint32_t nranlibs,
     enum NXByteOrder target_byte_order);
 
 extern void swap_relocation_info(
     struct relocation_info *relocs,
-    unsigned long nrelocs,
+    uint32_t nrelocs,
     enum NXByteOrder target_byte_order);
 
 extern void swap_indirect_symbols(
-    unsigned long *indirect_symbols,
-    unsigned long nindirect_symbols,
+    uint32_t *indirect_symbols,
+    uint32_t nindirect_symbols,
     enum NXByteOrder target_byte_sex);
 
 extern void swap_dylib_reference(
     struct dylib_reference *refs,  
-    unsigned long nrefs,
+    uint32_t nrefs,
     enum NXByteOrder target_byte_sex);
 
 extern void swap_dylib_module(  
     struct dylib_module *mods,
-    unsigned long nmods, 
+    uint32_t nmods, 
     enum NXByteOrder target_byte_sex);
 
 extern void swap_dylib_table_of_contents(
     struct dylib_table_of_contents *tocs,
-    unsigned long ntocs,
+    uint32_t ntocs,
     enum NXByteOrder target_byte_sex);
 
 #endif /* _MACH_O_SWAP_H_ */
